pyproject.toml


[project]
name = "traffic-api-data"
version = "0.0.1"
requires-python = ">= 3.12"
dependencies = [
    "python-decouple",
    "pytz",
    "quixstreams",
    "requests",
    "ruff"
]


[project.optional-dependencies]
dev = [
    "ipywidgets",
    "jupyterlab"
]


